+++ js/api.js
@@ -67,6 +67,14 @@ define(["dao", "globals", "ui", "core/freeAgents", "core/game", "core/league", "
             if (g.phase === g.PHASE.PRESEASON) {
                 phase.newPhase(g.PHASE.REGULAR_SEASON);
             }
+        } else if (amount === "stopAutoPlay") {
+            league.setGameAttributesComplete({autoPlaySeasons: 0}).then(function () {
+                ui.updatePlayMenu();
+                play("stop");
+
+                // Extra toggle to counteract play("stop");
+                $("#play-menu .dropdown-toggle").dropdown("toggle");
+            });
         }
 
         // Close the menu