+++ js/core/phase.js
@@ -30,6 +30,14 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
                 updateEvents.push("newPhase");
                 ui.realtimeUpdate(updateEvents, url);
             });
+        }).then(function () {
+            // If auto-simulating, initiate next action
+            if (g.autoPlaySeasons > 0) {
+                // Not totally sure why setTimeout is needed, but why not?
+                setTimeout(function () {
+                    require("core/league").autoPlay();
+                }, 100);
+            }
         });
     }
 
@@ -81,6 +89,10 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
                     }
                 });
             }).then(function () {
+                if (g.autoPlaySeasons > 0) {
+                    return require("core/league").setGameAttributes(tx, {autoPlaySeasons: g.autoPlaySeasons - 1});
+                }
+            }).then(function () {
                 if (g.enableLogging && !window.inCordova) {
                     ads.show();
                 }
@@ -128,7 +140,7 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
                 });
             }
             if ((localStorage.nagged === "2" && Math.random() < 0.25) || (localStorage.nagged === "3" && Math.random < 0.025)) {
-                _gaq.push(["_trackEvent", "Ad Display", "DraftKings"]);
+                if (g.enableLogging) { _gaq.push(["_trackEvent", "Ad Display", "DraftKings"]); }
                 localStorage.nagged = "3";
                 return dao.messages.add({
                     ot: tx,
@@ -510,7 +522,7 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
                 index: "tid",
                 key: IDBKeyRange.lowerBound(0),
                 callback: function (p) {
-                    if (p.contract.exp <= g.season && p.tid === g.userTid) {
+                    if (p.contract.exp <= g.season && p.tid === g.userTid && g.autoPlaySeasons === 0) {
                         // Add to free agents first, to generate a contract demand
                         return player.addToFreeAgents(tx, p, g.PHASE.RESIGN_PLAYERS, baseMoods).then(function () {
                             // Open negotiations with player
@@ -530,8 +542,8 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
             });
         }).then(function () {
             // Set daysLeft here because this is "basically" free agency, so some functions based on daysLeft need to treat it that way (such as the trade AI being more reluctant)
-            require("core/league").setGameAttributes(null, {daysLeft: 30});
-
+            return require("core/league").setGameAttributes(tx, {daysLeft: 30});
+        }).then(function () {
             return [helpers.leagueUrl(["negotiation"]), ["playerMovement"]];
         });
     }
@@ -568,7 +580,7 @@ define(["dao", "globals", "ui", "core/contractNegotiation", "core/draft", "core/
                         callback: function (p) {
                             var contract, factor;
 
-                            if (p.contract.exp <= g.season && p.tid !== g.userTid) {
+                            if (p.contract.exp <= g.season && (p.tid !== g.userTid || g.autoPlaySeasons > 0)) {
                                 // Automatically negotiate with teams
                                 if (strategies[p.tid] === "rebuilding") {
                                     factor = 0.4;