+++ js/util/ads.js
@@ -3,7 +3,7 @@
  * @namespace Everyone loves advertisements, right?
  */
 /*eslint new-cap: 0*/
-define([], function () {
+define(["globals"], function (g) {
     "use strict";
 
     function showGCS() {
@@ -39,6 +39,11 @@ define([], function () {
 
     function show() {
         var adTimer, now;
+
+        // No ads during multi season auto sim
+        if (g.autoPlaySeasons > 0) {
+            return;
+        }
 
         if (Math.random() < 0.75) {
             showGCS();