+++ CHANGES.md
@@ -9,6 +9,7 @@ number or not.
 - Export Stats feature (player average stats or individual game stats)
 - More detailed info display: four factors in box scores, transaction log in player pages, etc.
 - Statistical Feats feature
+- Auto play multiple seasons
 
 ## Techincal details